move kafka files to other dir
